smol_instruments/pyproject.toml
[project]
name = "smol-instruments"
version = "1.0.0"
description = "Production-ready smolagent ReAct for low-power LLMs"
requires-python = ">=3.10"
dependencies = [
    "smolagents>=0.3.0",
    "litellm>=1.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "openinference-instrumentation-smolagents>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
agent = "agent_runtime.run:main"

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
